requirements.txt
fastapi
uvicorn
Pillow
pandas
numpy
tensorflow
joblib
pydantic
python-multipart
